Fix the publisher to compare proper types.
I put strings in the scheduled but the IDs from DEV are integers.
This commit fixed the schedule and fixed the checking code
so that the comparison should always work in the future.
This should fix the periodic rate limiting because the periodic check
will only publish when there's something to publish.