Update 2020-12-16-middleware-do-you-go.md
On lines 352/425/499 I proposed a file name `project\middleware.py`. I’m not sure if this is the right path, but I think that since you’ve got a file name in the `settings.py` file it makes sense to call it out there. 

I also updated the Django docs link from 3.1 to 4.1